+++ vscode/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "note-searcher" extension will be documented in this
 file. Format inspired by https://keepachangelog.com/en/1.0.0/
 
 
+## [Unreleased]
+### Change
+- Dead links are now shown in the sidebar. Config no longer necessary - this
+  feature is always on.
+
 ## [0.0.16] - 2020-06-10
 ### Added
 - Show dead links on save (enable in preferences - off by default)